requirements.txt
cachetools==5.3.3
google-auth==2.31.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyfiglet==1.0.2
requests-oauthlib==2.0.0
rsa==4.9
